Cleanup; tests; fixes examples, geodesic, plotting
