+++ src/einsteinpy/examples.py
@@ -6,10 +6,9 @@ from einsteinpy.metric import Schwarzschild
 
 
 def perihelion():
-    # ToDo: Use dedicated coordinate class, after refactoring `coordinate`
     M = 6e24  # Mass
     t = 10000  # Coordinate Time (has no effect in this case - Schwarzschild)
-    x_vec = np.array([130.0, 0.0, 0.0])  # 3-Pos
+    x_vec = np.array([130.0, np.pi / 2, -np.pi / 8])  # 3-Pos
     v_vec = np.array([0.0, 0.0, 1900.0])  # 3-Vel
 
     # Schwarzschild Metric Object
@@ -18,7 +17,7 @@ def perihelion():
     x_4vec = four_position(t, x_vec)
     # Calculating Schwarzschild Metric at x_4vec
     ms_cov_mat = ms_cov.metric_covariant(x_4vec)
-    # Getting stacked initial vector
+    # Getting stacked (Length-8) initial vector, containing 4-Pos and 4-Vel
     init_vec = stacked_vec(ms_cov_mat, t, x_vec, v_vec, time_like=True)
 
     # Calculating Geodesic
