+++ src/einsteinpy/metric/__init__.py
@@ -1,4 +1,4 @@
-from .generic_metric import GenericMetric
+from .base_metric import BaseMetric
 from .kerr import Kerr
 from .kerrnewman import KerrNewman
 from .schwarzschild import Schwarzschild
