+++ tests/test_metric/test_schwarzschild.py
@@ -7,7 +7,6 @@ def test_schwarzschild_private_attr():
     """
     Tries to invoke private methods of Kerr
 
-    charge_geometrized = _private
     em_potential_covariant = _private
     em_potential_contravariant = _private
     em_tensor_covariant = _private
@@ -17,7 +16,6 @@ def test_schwarzschild_private_attr():
     obj = Schwarzschild(M=6e24)
 
     try:
-        obj.charge_geometrized()
         obj.em_tensor_covariant()
         obj.em_tensor_contravariant()
         obj.em_potential_covariant()
